voxcraft/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
sse-starlette>=2.2.0
pydantic-settings>=2.7.0
aiofiles>=24.1.0
numpy>=1.26.0
soundfile>=0.12.0
openai>=1.0.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
ebooklib>=0.18
psutil>=5.9.0
httpx>=0.27.0
# mlx-audio is Apple Silicon only - install locally, not in Docker
